requirements.txt
absl-py==0.9.0
astunparse==1.6.3
autopep8==1.5.2
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
cycler==0.10.0
Flask==1.1.2
Flask-WTF==0.14.3
gast==0.3.3
google-auth==1.15.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.29.0
gunicorn==20.0.4
h5py==2.10.0
idna==2.9
imageio==2.8.0
importlib-metadata==1.6.0
itsdangerous==1.1.0
Jinja2==2.11.2
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.2.1
numpy==1.18.4
oauthlib==3.1.0
opt-einsum==3.2.1
Pillow==7.1.2
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.13.0
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
scipy==1.4.1
six==1.15.0
tensorboard==2.2.1
tensorboard-plugin-wit==1.6.0.post3
tensorflow==2.2.0
tensorflow-estimator==2.2.0
termcolor==1.1.0
urllib3==1.25.9
Werkzeug==1.0.1
wrapt==1.12.1
WTForms==2.3.1
zipp==3.1.0
